Improving logic, renaming module
